+++ Matriz/kCrossVal.cpp
@@ -26,18 +26,18 @@ int main(int argc, char* argv[]){
 		int k = atoi(argv[7]);
 
 		int tamConj = cantIm/k;
-		vector<float> hitRate;
+		vector<double> hitRate;
 		for(int i = 0; i<k; i++){
 			Matriz test = imagenes.subMatriz(i*tamConj,(i+1)*tamConj);
 			Matriz train = imagenes.subMatriz(0,i*tamConj,(i+1)*tamConj,cantIm);
-			float hR = usarPca(train,test,cantAutov,cantIterMetPot,cantVecinos);
+			double hR = usarPca(train,test,cantAutov,cantIterMetPot,cantVecinos);
 			hitRate.push_back(hR);
 		}
-		float hR = 0;
+		double hR = 0;
 		for (int i = 0; i < hitRate.size(); ++i){
 			hR += hitRate[i];
 		}
-		hR = hR/(float) hitRate.size();
+		hR = hR/(double) hitRate.size();
 		cout << "HitRate: " << hR << endl;	
 	} else if (strcmp(argv[1],"pls") == 0){
 		int cantIm = atoi(argv[3]);
@@ -49,18 +49,18 @@ int main(int argc, char* argv[]){
 		int k = atoi(argv[7]);
 
 		int tamConj = cantIm/k;
-		vector<float> hitRate;
+		vector<double> hitRate;
 		for(int i = 0; i<k; i++){
 			Matriz test = imagenes.subMatriz(i*tamConj,(i+1)*tamConj);
 			Matriz train = imagenes.subMatriz(0,i*tamConj,(i+1)*tamConj,cantIm);
-			float hR = usarPls(train,test,cantIterPls,cantIterMetPot,cantVecinos);
+			double hR = usarPls(train,test,cantIterPls,cantIterMetPot,cantVecinos);
 			hitRate.push_back(hR);
 		}
-		float hR = 0;
+		double hR = 0;
 		for (int i = 0; i < hitRate.size(); ++i){
 			hR += hitRate[i];
 		}
-		hR = hR/(float) hitRate.size();
+		hR = hR/(double) hitRate.size();
 		cout << "HitRate: " << hR << endl;	
 	} else if (strcmp(argv[1],"knn") == 0){
 		int cantIm = atoi(argv[3]);
@@ -70,18 +70,18 @@ int main(int argc, char* argv[]){
 		int k = atoi(argv[5]);
 
 		int tamConj = cantIm/k;
-		vector<float> hitRate;
+		vector<double> hitRate;
 		for(int i = 0; i<k; i++){
 			Matriz test = imagenes.subMatriz(i*tamConj,(i+1)*tamConj);
 			Matriz train = imagenes.subMatriz(0,i*tamConj,(i+1)*tamConj,cantIm);
-			float hR = usarKnn(train,test,cantVecinos);
+			double hR = usarKnn(train,test,cantVecinos);
 			hitRate.push_back(hR);
 		}
-		float hR = 0;
+		double hR = 0;
 		for (int i = 0; i < hitRate.size(); ++i){
 			hR += hitRate[i];
 		}
-		hR = hR/(float) hitRate.size();
+		hR = hR/(double) hitRate.size();
 		cout << "HitRate: " << hR << endl;
 	}
