+++ Matriz/parser.cpp
@@ -11,7 +11,7 @@ Matriz parser(char* file, int cantImagenes){ // cantImagenes con -1 es equivalen
 	
 	ifstream in(file);
 	if (in.is_open()){
-		vector< vector <float> > mtx;
+		vector< vector <double> > mtx;
 		vector<int> numeroRepresentado;
 
 		// Parseo de la primer línea
@@ -34,7 +34,7 @@ Matriz parser(char* file, int cantImagenes){ // cantImagenes con -1 es equivalen
 			char aux;
 			iss>> aux;
 			//Aca extraemos la info de la línea en la que estamos
-			vector<float> img;
+			vector<double> img;
 			for (int i = 0; i<784; i++){
 
 				int pixel;
@@ -62,7 +62,7 @@ Matriz parserTest2(char* file, int cantImagenes){ // cantImagenes con -1 es equi
 	
 	ifstream in(file);
 	if (in.is_open()){
-		vector< vector <float> > mtx;
+		vector< vector <double> > mtx;
 		vector<int> numeroRepresentado;
 
 		// Parseo de la primer línea
@@ -79,7 +79,7 @@ Matriz parserTest2(char* file, int cantImagenes){ // cantImagenes con -1 es equi
 			iss.str(st);
 
 			//Aca extraemos la info de la línea en la que estamos
-			vector<float> img;
+			vector<double> img;
 			for (int i = 0; i<784; i++){
 
 				int pixel;
@@ -111,7 +111,7 @@ Matriz parserImgTest(char* file, int imgDesde, int imgHasta){
 	
 	ifstream in(file);
 	if (in.is_open()){
-		vector< vector <float> > mtx;
+		vector< vector <double> > mtx;
 		vector<int> numeroRepresentado;
 
 		// Parseo de la primer línea
@@ -138,7 +138,7 @@ Matriz parserImgTest(char* file, int imgDesde, int imgHasta){
 			char aux;
 			iss>> aux;
 			//Aca extraemos la info de la línea en la que estamos
-			vector<float> img;
+			vector<double> img;
 			for (int i = 0; i<784; i++){
 
 				int pixel;
